+++ controllers/employeeController.js
@@ -3,8 +3,13 @@ const Employee = require("../models/employee");
 
 const employeeGet = async (req, res) => {
   try {
-    const employee = await Employee.find({});
-    res.send(employee);
+    if (req.isAuthenticated() && req.user.role === "admin"){
+      const employee = await Employee.find({});
+      res.send(employee);
+    } else {
+      res.send("Restricted access");
+    }
+
   }
   catch (err) {
     console.log(err);
@@ -17,8 +22,12 @@ const employeePut = async (req, res) => {
 
 const employeePost = async (req, res) => {
   try {
-    const employee = await Employee.create(req.body);
-    res.send(employee);
+    if (req.isAuthenticated() && req.user.role === "admin") {
+      const employee = await Employee.create(req.body);
+      res.send(employee);
+    } else {
+      res.send("Restricted access");
+    }
   }
   catch (err) {
     console.log(err);
@@ -31,8 +40,12 @@ const employeeDelete = async (req, res) => {
 
 const particularEmpGet = async (req, res) => {
   try {
-    const employee = await Employee.findById(req.params.employeeId);
-    res.send(employee);
+    if (req.isAuthenticated()) {
+      const employee = await Employee.findById(req.params.employeeId);
+      res.send(employee);
+    } else {
+      res.send("Restricted access");
+    }
   }
   catch (err) {
     console.log(err);
@@ -45,12 +58,16 @@ const particularEmpPost = async (req, res) => {
 
 const particularEmpPut = async (req, res) => {
   try {
-    const employee = await Employee.findByIdAndUpdate(
-      req.params.employeeId,
-      {$set: req.body},
-      {new: true}
-    );
-    res.send(employee);
+    if (req.isAuthenticated() && req.user.role !== "customer") {
+      const employee = await Employee.findByIdAndUpdate(
+        req.params.employeeId,
+        {$set: req.body},
+        {new: true}
+      );
+      res.send(employee);
+    } else {
+      res.send("Restricted access");
+    }
   }
   catch (err) {
     console.log(err);
@@ -59,8 +76,12 @@ const particularEmpPut = async (req, res) => {
 
 const particularEmpDelete = async (req, res) => {
   try {
-    const employee = await Employee.findByIdAndRemove(req.params.employeeId);
-    res.send("Employee successfully removed");
+    if (req.isAuthenticated() && req.user.role === "admin") {
+      const employee = await Employee.findByIdAndRemove(req.params.employeeId);
+      res.send("Employee successfully removed");
+    } else {
+      res.send("Restricted access");
+    }
   }
   catch (err) {
     console.log(err);
